requirements.txt
networkx
matplotlib
biopython
pygraphviz
pydot
